Add files via upload
To overcome the problem of traffic congestion, the traffic prediction using
machine learning which contains regression model and libraries like pandas, os,
numpy, matplotlib.pyplot are used to predict the traffic. This has to be
implemented so that the traffic congestion is controlled and can be accessed
easily. Users can collect the traffic information of the traffic flow and can also
check the congestion flow from the start of the day till the end of the day with the
time span of one hour data. In this way, Users can know the weather conditions of
the roads that they would probably opt to take. This also tells the accuracy of the
traffic by comparing their mean square errors of the past year’s data and the recent
year’s data. Users can also know how many vehicles are traveling on average by
the traffic prediction.